split VecGraph bindings into their own file
